+++ nix/modules/project/defaults.nix
@@ -84,6 +84,11 @@ in
       '';
       default =
         let
+          globalSettings = {
+            # We disable this by default because it causes breakage.
+            # See https://github.com/srid/haskell-flake/pull/253
+            buildFromSdist = lib.mkDefault false;
+          };
           localSettings = { name, package, config, ... }:
             lib.optionalAttrs (package.local.toDefinedProject or false) {
               # Disabling haddock and profiling is mainly to speed up Nix builds.
@@ -91,7 +96,12 @@ in
               libraryProfiling = lib.mkDefault false; # Avoid double-compilation.
             };
         in
-        if config.defaults.enable then localSettings else { };
+        if config.defaults.enable then {
+          imports = [
+            globalSettings
+            localSettings
+          ];
+        } else { };
     };
 
     projectModules.output = mkOption {